refactor: remove bars/runtime shims
